fim da lapidacao do componente
