+++ README.md
@@ -18,3 +18,9 @@ Como minimizar dependências de um Componente
 A diretiva \*ngTemplateOutlet
 ComponentFactoryResolver para criar fábricas de Componentes
 O papel de ApplicationRef e HotView de um ComponentRef
+
+#
+
+Estilização padrão de um modal
+Animações com o módulo Animations
+Como reutilizar animações
