+++ src/app/app.component.ts
@@ -1,16 +1,21 @@
 import { Component, TemplateRef, ViewChild } from '@angular/core';
-import { ModalRef, ModalService } from './shared/components/modal/services/modal.service';
+
+import { fade } from './shared/animations/fade';
+import { ModalRef } from './shared/components/modal/models/modal-ref';
+import { ModalService } from './shared/components/modal/services/modal.service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  styleUrls: ['./app.component.scss'],
+  animations: [fade]
 })
 export class AppComponent {
   @ViewChild('modal') public modalTemplateRef: TemplateRef<any>;
   title = 'a11y-p2';
   public firstName = 'Flávio';
   public modalRef: ModalRef;
+  public info = false;
 
   constructor(private modalService: ModalService) {}
