+++ src/app/shared/components/modal/modal.component.html
@@ -1,5 +1,9 @@
-<div class="modal-overlay">
-  <div class="modal-content">
+<div
+  class="modal-overlay"
+  (click)="modalRef.close()"
+  >
+  <div class="modal-content"
+    (click)="$event.stopPropagation()">
     <h1 class="modal-title">{{ config.title }}</h1>
     <ng-container *ngTemplateOutlet="config.templateRef"></ng-container>
   </div>
