+++ src/app/shared/components/modal/models/modal-ref.ts
@@ -1,0 +1,12 @@
+import { ComponentRef } from '@angular/core';
+import { ModalComponent } from '../modal.component';
+
+export class ModalRef {
+
+  constructor(private componentRef: ComponentRef<ModalComponent>) {}
+
+  public close(): void {
+    console.log('close called');
+    this.componentRef.destroy();
+  }
+}
