+++ src/app/shared/components/modal/services/modal.service.ts
@@ -1,7 +1,9 @@
 import { ComponentFactory, ComponentFactoryResolver, ComponentRef, Injectable, Injector } from '@angular/core';
+
 import { BodyInjectorService } from 'src/app/shared/services/body-injector';
 import { ModalConfig } from '../interfaces/modal-config';
 import { ModalComponent } from '../modal.component';
+import { ModalRef } from '../models/modal-ref';
 
 @Injectable()
 export class ModalService {
@@ -22,20 +24,12 @@ export class ModalService {
     console.log(componentRef.instance);
     console.log('open called');
     this.bodyInjector.stackBeforeAppRoot(componentRef);
-    return new ModalRef(componentRef);
+    const modalRef = new ModalRef(componentRef);
+    componentRef.instance.modalRef = modalRef;
+    return modalRef;
   }
 
   private createComponentRef(): ComponentRef<ModalComponent> {
     return this.componentFactory.create(this.injector);
-  }
-}
-
-export class ModalRef {
-
-  constructor(private componentRef: ComponentRef<ModalComponent>) {}
-
-  public close(): void {
-    console.log('close called');
-    this.componentRef.destroy();
   }
 }
